Cleanup and harden Enum code against open generic enums.
- Phase out the use of EETypePtr.IsEnum due to

  https://github.com/dotnet/corert/issues/3802

- Remove the no-longer needed split between
  EnumInfo and EnumInfoImplementation.

- Unbundle the IsEnum check from GetEnumInfo()
  so that validation order can be more easily
  compared with CoreCLR code.

- Harden against open generic enums:

  - Fall back to metadata to get the underlying
    CorElementInfo.

  - Make sure apis that instantiate the enum
    type (e.g. Enum.ToObject()) throw a well-defined
    exception if given an open generic enum.

    https://github.com/dotnet/coreclr/issues/11299

    tracks a similar issue on CoreCLR (which
    dementedly creates an instanced object
    of an open generic type...)

One has to wonder why the policy of wishing
a type's generic parameters onto its nested
types couldn't have had an exemption for enums...
