+++ src/System.Private.CoreLib/src/Resources/Strings.resx
@@ -2446,4 +2446,7 @@
   <data name="Serialization_DelegatesNotSupported" xml:space="preserve">
     <value>Serializing delegates is not supported on this platform.</value>
   </data>
+  <data name="Arg_OpenType" xml:space="preserve">
+    <value>Cannot create an instance of {0} as it is an open type.</value>
+  </data>
 </root>
